Database rewrite - added GetItem by id, GetItems per purchase and per Shop, cleanup - if no database exists in appdata, one is copied from the assembly - rewrote Item, ItemPurchaseData and ItemPurchaseHistory to fit the database needs - currency now represented as int - StatusBarViewModel cleanup

+++ Core/Serialization/ItemPurchaseHistory.cs
@@ -16,14 +16,14 @@ namespace BillScannerCore {
 		/// <summary>
 		/// How many items of this type were bought at that time
 		/// </summary>
-		public long amount { get; }
+		public int amount { get; }
 
 		/// <summary>
 		/// The amount of money the item cost at that time
 		/// </summary>
 		public int price { get; }
 
-		public ItemPurchaseHistory(int purchaseID, long amount, int price) {
+		public ItemPurchaseHistory(int purchaseID, int amount, int price) {
 			this.purchaseID = purchaseID;
 			this.amount = amount;
 			this.price = price;
